fix timing print bug
caused by camera error correction. Even if the game has changed the player's turn and the player is unable to move to avoid calling the focus of the camera. The time continued to count for the current player. Now the time remains intact until the previous player completes the move and the current player becomes able to play.